Feature/XAFlagsFromXAResource

* nuke flags

* sonar

+++ casual/casual-api/src/main/java/se/kodarkatten/casual/api/flags/AtmiFlags.java
@@ -28,7 +28,7 @@ public enum AtmiFlags implements CasualFlag
     }
 
     @Override
-    public final long getValue() {
+    public final int getValue() {
         return value;
     }
