+++ casual/casual-network/src/main/java/se/kodarkatten/casual/network/messages/transaction/CasualTransactionResourcePrepareRequestMessage.java
@@ -13,7 +13,7 @@ import java.util.UUID;
  */
 public final class CasualTransactionResourcePrepareRequestMessage extends AbstractCasualTransactionRequestMessage
 {
-    private CasualTransactionResourcePrepareRequestMessage(final UUID execution, final Xid xid, long resourceId, long flags)
+    private CasualTransactionResourcePrepareRequestMessage(final UUID execution, final Xid xid, long resourceId, int flags)
     {
         super(execution, xid, resourceId, flags);
     }
